requirements.txt
# Core ML & Computer Vision
ultralytics>=8.3.0
opencv-python>=4.8.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
pandas>=2.0.0
scikit-image>=0.21.0

# Geometry & Spatial
shapely>=2.0.0

# Visualization & Reporting
matplotlib>=3.7.0
seaborn>=0.12.0
reportlab>=4.0.0
Pillow>=10.0.0

# API & Web
fastapi>=0.100.0
uvicorn>=0.23.0
requests>=2.31.0

# Configuration & Environment
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Tracking (if needed separately)
# lap>=0.4.0  # for ByteTrack
